# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beyond Plus CMS API Documentation",
        "description": "JSON API for Beyond Plus CMS.",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/m/posts/{slug}": {
            "get": {
                "tags": [
                    "CMS"
                ],
                "summary": "Single published post by slug",
                "operationId": "8f8393851f8bf7939d4986acac3f4c7b",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/m/pages/{slug}": {
            "get": {
                "tags": [
                    "CMS"
                ],
                "summary": "Single published page by slug",
                "operationId": "e163a65b3e0cc160ab64da61bd21a843",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/m/categories/{slug}/posts": {
            "get": {
                "tags": [
                    "CMS"
                ],
                "summary": "Published posts in a category",
                "operationId": "831b70046832033cd4cfb306ce9ee560",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/m/search": {
            "get": {
                "tags": [
                    "CMS"
                ],
                "summary": "Search published posts and pages by title or body",
                "operationId": "7dfee178430977406403cc52e399de24",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}